+++ src/controllers/cardController.ts
@@ -1,24 +1,28 @@
+import { AuthRequest } from "../middlewares/authMiddleware";
 import { CardService } from "../services/cardService";
 import { Request, Response } from 'express';
 
 export class CardController{
     private service: CardService;
 
-    constructor(){
-        this.service = new CardService();
+    constructor(service:CardService){
+        this.service = service;
     }
 
-    async create(req:Request,res:Response):Promise<void>{
+    async create(req:AuthRequest,res:Response):Promise<Response>{
         try{
             const {ownerId,collectionCode,number,modifier} = req.body;
+            if (req.user.id != ownerId) {
+                return res.status(400).json({message:'Sem permissão'})
+            }
             if (!ownerId || !collectionCode || !number || !modifier){
-                res.status(400).json({message:'Todos os campos são obrigatórios.'});
+                return res.status(400).json({message:'Todos os campos são obrigatórios.'});
             }
             const new_ = await this.service.create({ownerId,collectionCode,number,modifier});
-            res.status(201).json(new_);
+            return res.status(201).json(new_);
 
         } catch (error){
-            res.status(500).json({ message: (error instanceof Error ? error.message : 'Internal server error') });
+            return res.status(500).json({ message: (error instanceof Error ? error.message : 'Internal server error') });
         }  
     }
 
@@ -35,34 +39,40 @@ export class CardController{
         }  
     }
 
-    async update(req:Request,res:Response):Promise<void>{
+    async update(req:AuthRequest,res:Response):Promise<Response>{
         try{
             const id = Number(req.query.id);
             if (!id){
-                res.status(400).json({message:'Id não informado.'});
+                return res.status(400).json({message:'Id não informado.'});
+            }
+            if (req.user.id != (await this.service.get(id)).ownerId) {
+                return res.status(400).json({message:'Sem permissão'})
             }
             const {ownerId,collectionCode,number,modifier} = req.body;
             if(!ownerId && !collectionCode && !number && !modifier){
-                res.status(400).json({message:'sem informações para atualizar'});
+                return res.status(400).json({message:'sem informações para atualizar'});
             }
             const new_ = await this.service.update(id,{ownerId,collectionCode,number,modifier});
-            res.status(201).json(new_)
+            return res.status(201).json(new_)
 
         } catch (error){
-            res.status(500).json({ message: (error instanceof Error ? error.message : 'Internal server error') });
+            return res.status(500).json({ message: (error instanceof Error ? error.message : 'Internal server error') });
         }  
     }
 
-    async delete(req:Request,res:Response):Promise<void>{
+    async delete(req:AuthRequest,res:Response):Promise<Response>{
         try{
             const id = Number(req.query.id);
             if (!id){
-                res.status(400).json({message:'Id não informado.'});
+                return res.status(400).json({message:'Id não informado.'});
+            }
+            if (req.user.id != (await this.service.get(id)).ownerId) {
+                return res.status(400).json({message:'Sem permissão'})
             }
             const new_ = await this.service.delete(id);
-            res.status(201).json(new_)
+            return res.status(201).json(new_)
         } catch (error){
-            res.status(500).json({ message: (error instanceof Error ? error.message : 'Internal server error') });
+            return res.status(500).json({ message: (error instanceof Error ? error.message : 'Internal server error') });
         }  
     }
 }
