+++ src/controllers/deckController.ts
@@ -1,24 +1,31 @@
+import { AuthRequest } from "../middlewares/authMiddleware";
+import { CardService } from "../services/cardService";
 import { DeckService } from "../services/deckService";
 import { Request, Response } from 'express';
 
 export class DeckController{
     private service: DeckService;
+    private cardService:CardService;
 
-    constructor(){
-        this.service = new DeckService();
+    constructor(service:DeckService,cardService:CardService){
+        this.service = service;
+        this.cardService = cardService;
     }
 
-    async create(req:Request,res:Response):Promise<void>{
+    async create(req:AuthRequest,res:Response):Promise<Response>{
         try{
             const {ownerId,name,victories,losses} = req.body;
+            if (req.user.id != ownerId) {
+                return res.status(400).json({message:'Sem permissão'})
+            }
             if (!ownerId || !name){
-                res.status(400).json({message:'nome e dono são obrigatórios'});
+                return res.status(400).json({message:'nome e dono são obrigatórios'});
             }
             const new_ = await this.service.create({ownerId,name,victories,losses});
-            res.status(201).json(new_);
+            return res.status(201).json(new_);
 
         } catch (error){
-            res.status(500).json({ message: (error instanceof Error ? error.message : 'Internal server error') });
+            return res.status(500).json({ message: (error instanceof Error ? error.message : 'Internal server error') });
         }  
     }
 
@@ -35,64 +42,82 @@ export class DeckController{
         }  
     }
 
-    async update(req:Request,res:Response):Promise<void>{
+    async update(req:AuthRequest,res:Response):Promise<Response>{
         try{
             const id = Number(req.query.id);
             if (!id){
-                res.status(400).json({message:'Id não informado.'});
+                return res.status(400).json({message:'Id não informado.'});
+            }
+            if (req.user.id != (await this.service.get(id)).ownerId) {
+                return res.status(400).json({message:'Sem permissão'})
             }
             const {ownerId,name,victories,losses} = req.body;
             if(!ownerId && !name && !victories && !losses){
-                res.status(400).json({message:'sem informações para atualizar'});
+                return res.status(400).json({message:'sem informações para atualizar'});
             }
             const new_ = await this.service.update(id,{ownerId,name,victories,losses});
-            res.status(201).json(new_)
+            return res.status(201).json(new_)
 
         } catch (error){
-            res.status(500).json({ message: (error instanceof Error ? error.message : 'Internal server error') });
+            return res.status(500).json({ message: (error instanceof Error ? error.message : 'Internal server error') });
         }  
     }
 
-    async delete(req:Request,res:Response):Promise<void>{
+    async delete(req:AuthRequest,res:Response):Promise<Response>{
         try{
             const id = Number(req.query.id);
             if (!id){
-                res.status(400).json({message:'Id não informado.'});
+                return res.status(400).json({message:'Id não informado.'});
+            }
+            if (req.user.id != (await this.service.get(id)).ownerId) {
+                return res.status(400).json({message:'Sem permissão'})
             }
             const new_ = await this.service.delete(id);
-            res.status(201).json(new_)
+            return res.status(201).json(new_)
         } catch (error){
-            res.status(500).json({ message: (error instanceof Error ? error.message : 'Internal server error') });
+            return res.status(500).json({ message: (error instanceof Error ? error.message : 'Internal server error') });
         }  
     }
 
-    async add_card(req:Request,res:Response):Promise<void>{
+    async add_card(req:AuthRequest,res:Response):Promise<Response>{
         try{
             const {deckId,cardId} = req.body;
+            if (req.user.id != (await this.service.get(deckId)).ownerId) {
+                return res.status(400).json({message:'Sem permissão'})
+            }
+            if (req.user.id != (await this.cardService.get(cardId)).ownerId) {
+                return res.status(400).json({message:'Sem permissão'})
+            }
             await this.service.add_card(deckId,cardId);
-             res.status(201).json({message:"success"})
+            return res.status(201).json({message:"success"})
         } catch (error){
-            res.status(500).json({ message: (error instanceof Error ? error.message : 'Internal server error') });
+            return res.status(500).json({ message: (error instanceof Error ? error.message : 'Internal server error') });
         }  
     }
 
-    async remove_card(req:Request,res:Response):Promise<void>{
+    async remove_card(req:AuthRequest,res:Response):Promise<Response>{
         try {
             const {deckId,cardId} = req.body;
+            if (req.user.id != (await this.service.get(deckId)).ownerId) {
+                return res.status(400).json({message:'Sem permissão'})
+            }
+            if (req.user.id != (await this.cardService.get(cardId)).ownerId) {
+                return res.status(400).json({message:'Sem permissão'})
+            }
             await this.service.remove_card(deckId,cardId);
-             res.status(201).json({message:"success"})
+            return res.status(201).json({message:"success"})
         } catch (error){
-            res.status(500).json({ message: (error instanceof Error ? error.message : 'Internal server error') });
+            return res.status(500).json({ message: (error instanceof Error ? error.message : 'Internal server error') });
         }  
     }
     
-    async get_cards(req:Request,res:Response):Promise<void>{
+    async get_cards(req:Request,res:Response):Promise<Response>{
         try {
             const id = Number(req.query.id);
             const _cards = await this.service.get_cards(id);
-             res.status(201).json(_cards)
+            return res.status(201).json(_cards)
         } catch (error){
-            res.status(500).json({ message: (error instanceof Error ? error.message : 'Internal server error') });
+            return res.status(500).json({ message: (error instanceof Error ? error.message : 'Internal server error') });
         }  
     }
 }
